Market Screen
Added the market screen (minus the scrollview with the actual goods, getting that done after finishing planet screen, which is next). Made it accessible from the planet screen (button works). Had to comment out the testing of the Good classes in PlayerConfirmation.java, as there was an error when creating the Planet