Market Screen Initial Values
Changed player in market class to reference the singleton. Added a method in Ship to get the free space. Market Screen will now initialize all values it needs to initialize from the planet and the player, and the screen is complete. All that is left is button functionality and updating values as the player buys things. Market goods also disappear at creation, as they are intended to be revealed when the player enter buy mode or sell mode. I'm doing that next.