BuyGood works!
Edited the buyItem function of the Market object (so as to not calculatePrice every time). Made updateHoldQuantity to update the amount per resource in the player's ship. Buying goods works.
TODO: sellMode, showing the cost per good while inputting quantities (before attempting the transaction), and fixing having a blank input for an item's quantity (currently that crashes the game)