Refuel
Refuel now works on the planet screen, and successfully subtracts the price, and toasts the price subtracted to the player on successful refuel. Currently, The price per gallon is equal to the game difficulty price modifier (base price per gallon is 1). Also removed some unused code, and changed most instances of getUniverse().get(0).getPlanetList().get(0) to getCurrentPlanet(). Left Player Confirmation functionally the same, to where it sets the current planet and solar systems to the 0th and 0th respectively.